requirements.txt
ago==0.0.93
appnope==0.1.2
attrs==20.3.0
Automat==20.2.0
awscli==1.19.53
backcall==0.2.0
beautifulsoup4==4.9.3
black==21.9b0
botocore==1.20.53
bs4==0.0.1
cchardet==2.1.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
debugpy==1.4.3
decorator==5.1.0
dnspython==2.1.0
docutils==0.15.2
dotmap==1.3.23
elasticsearch==7.12.0
entrypoints==0.3
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
h2==3.2.0
hjson==3.0.2
hpack==3.0.0
hurry.filesize==0.9
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
itemadapter==0.2.0
itemloaders==1.0.4
jedi==0.18.0
jieba3k==0.35.1
jmespath==0.10.0
joblib==1.0.1
jupyter-client==7.0.3
jupyter-core==4.8.1
langdetect==1.0.8
lxml==4.6.3
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.1
news-please==1.5.18
newspaper3k==0.2.8
nltk==3.6.1
numpy==1.20.2
oauthlib==3.1.1
pandas==1.2.4
parsel==1.6.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
plac==1.3.3
platformdirs==2.3.0
priority==1.3.0
prompt-toolkit==3.0.20
Protego==0.1.16
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.10.0
pymongo==3.11.3
PyMySQL==1.0.2
pyOpenSSL==20.0.1
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.3.0
queuelib==1.5.0
readability-lxml==0.8.1
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.7
Scrapy==2.5.0
service-identity==18.1.0
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2.1
tinysegmenter==0.3
tldextract==3.1.0
tomli==1.2.1
tornado==6.1
tqdm==4.60.0
traitlets==5.1.0
tweepy==3.10.0
Twisted==21.2.0
typing-extensions==3.10.0.2
urllib3==1.26.4
w3lib==1.22.0
warcio==1.7.4
wcwidth==0.2.5
zope.interface==5.4.0
